requirements.txt
Django==4.2.7
requests==2.31.0
python-decouple==3.8
celery==5.3.4